+++ Locales/enUS.lua
@@ -57,6 +57,15 @@ select(2, ...).L = setmetatable({
     ["SNIPPETS_TIPS"] = "[Double-Click] to rename. [Shift-Click] to delete. All checked snippets will be automatically invoked at the end of Cell initialization process (in ADDON_LOADED event).",
 
     ["CHANGELOGS"] = [[
+        <h1>r112-release (Sep 11, 2022, 19:00 GMT+8)</h1>
+        <p>* Add custom auras support to Defensives and Externals.</p>
+        <p>* Add Mirror Image to Defensives.</p>
+        <p>* Add Cell default texture to LibSharedMedia.</p>
+        <h2>Wrath Classic</h2>
+        <p>* Updated raid debuffs.</p>
+        <p>* Fixed power filter.</p>
+        <br/>
+
         <h1>r111-release (Sep 3, 2022, 12:07 GMT+8)</h1>
         <p>* Fixed game version check.</p>
         <p>* Updated zhTW.</p>