+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r110-release...04ec857ddda7ec23e832d55c0a0c264246417c01)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r111-release...1e4b9341b3ae0a1c0f81c9ff7b792f0795acf9a9)
 
-- Fix game version check
-- Update zhTW
+- Update WotLK raid debuffs
+- Add custom auras support to Defensives and Externals
+- Add Mirror Image to Defensives
+- Add Cell default texture to LibSharedMedia
+- Disable some status color options in WotLK
+- Fix power filter in WotLK
+- [snippet] RaidGroupNumber_Name